test: move the RedisClient async-context suite into its own file
